Added a template check
Some subfolders did not have a template. Added a check

+++ AU cluster/Greedy_MC_3Dstack.py
@@ -106,6 +106,7 @@ def MakeListAndHyperstack(tif_file_folder):
 
 tif_file_folder=sys.argv[1]
 tif_file_folder=os.path.normpath(tif_file_folder)
+print(tif_file_folder)
 os.chdir(os.path.dirname(tif_file_folder))
 img_seq_list=glob.glob(tif_file_folder+'/3Dreg/*.tif')
 template_name=tif_file_folder+'/3Dreg/template.tif'
@@ -118,23 +119,32 @@ MC_img_list=glob.glob(tif_file_folder+'/3Dreg/*'+os.path.basename(os.path.normpa
 MC_img_list=[x for x in MC_img_list if not 'LongReg' in x]
 print(tif_file_folder+' '+str(len(MC_img_list)))    
 f = open(tif_file_folder+'/ListOfFailedFiles.txt','w')
-if len(MC_img_list)==1200:
- C1_name=MC_img_list[0]
- file_name=os.path.basename(os.path.normpath(tif_file_folder))
+file_name=os.path.basename(os.path.normpath(tif_file_folder))
+if len(MC_img_list)==1200 and not os.path.exists(tif_file_folder+'/'+file_name+'_4D.tif'):
+ C1_name=MC_img_list[0] 
  range2=int(file_name.split('range')[-1].split('_')[0])
  step=int(file_name.split('step')[-1].split('_')[0])
  TrueSlices=int((range2/step)+1);    
- base_img=nib.load(C1_name)
+ try:
+  base_img=nib.load(C1_name)
+ except:
+  Register_single_image_forced(C1_name.replace('_Warped.nii',''),template_name,mask_name)
+  base_img=nib.load(C1_name)
  base_img=np.squeeze(np.asarray(base_img.get_fdata(),dtype='uint16')).transpose()        
  C1frames=np.zeros((int(len(MC_img_list)),TrueSlices,base_img.shape[1],base_img.shape[2]), dtype='uint16')
  for img_nb,C2_name in enumerate(MC_img_list):    
-  img_nb=int( re.search('_power\d+_(\d+)\.tif',C2_name).group(1))  
-  img_temp=nib.load(C2_name)
+  img_nb=int( re.search('_power\d+_(\d+)\.tif',C2_name).group(1)) 
+  try:
+   img_temp=nib.load(C2_name)
+  except:
+   Register_single_image_forced(C2_name.replace('_Warped.nii',''),template_name,mask_name)
+   img_temp=nib.load(C1_name)
   img_temp=img_temp.get_fdata()    
   img_temp=np.squeeze(np.asarray(img_temp,dtype='uint16')).transpose()
   C1frames[img_nb,:,:,:]=img_temp 
  tifffile.imwrite(tif_file_folder+'/'+file_name+'_4D.tif',C1frames)
  nrrd.write(tif_file_folder+'/'+file_name+'_4D.nrrd',C1frames)   
+ print(tif_file_folder + 'is done')
 else:
  test=[x.split('_Warped.nii')[0] for x in MC_img_list]
  test=set(test) ^ set(img_seq_list)
